requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
filelock==3.0.10
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
oauthlib==2.1.0
pluggy==0.8.0
py==1.7.0
PySocks==1.6.8
python-dotenv==0.10.1
requests==2.21.0
requests-oauthlib==1.0.0
six==1.12.0
toml==0.10.0
tox==3.6.1
tox-pip-extensions==1.4.1
tweepy==3.7.0
urllib3==1.24.1
virtualenv==16.1.0
Werkzeug==0.14.1
